move, rename and add test for component
