+++ app/src/organisms/OnDeviceDisplay/RobotDashboard/index.ts
@@ -1,0 +1,2 @@
+export * from './EmptyRecentRun'
+export * from './RecentProtocolRunCard'
