+++ app/src/pages/OnDeviceDisplay/RobotDashboard.tsx
@@ -1,29 +1,25 @@
 import * as React from 'react'
 import { useTranslation } from 'react-i18next'
-import { useHistory } from 'react-router-dom'
-import { formatDistance } from 'date-fns'
 
 import {
   Flex,
   DIRECTION_COLUMN,
   DIRECTION_ROW,
   SPACING,
-  COLORS,
   TYPOGRAPHY,
-  BORDERS,
 } from '@opentrons/components'
 import {
   useAllProtocolsQuery,
   useAllRunsQuery,
 } from '@opentrons/react-api-client'
-import { css } from 'styled-components'
 
 import { StyledText } from '../../atoms/text'
-import { Chip } from '../../atoms/Chip'
 import { Navigation } from '../../organisms/OnDeviceDisplay/Navigation'
 import { onDeviceDisplayRoutes } from '../../App/OnDeviceDisplayApp'
-import { EmptyRecentRun } from '../../organisms/OnDeviceDisplay/RobotDashboard/EmptyRecentRun'
-import { useMissingProtocolHardware } from '../Protocols/hooks'
+import {
+  EmptyRecentRun,
+  RecentProtocolRunCard,
+} from '../../organisms/OnDeviceDisplay/RobotDashboard'
 import { sortProtocols } from './ProtocolDashboard/utils'
 
 export const MAXIMUM_RECENT_RUN_PROTOCOLS = 8 // This might be changed
@@ -66,7 +62,7 @@ export function RobotDashboard(): JSX.Element {
                 )?.createdAt
                 return (
                   <React.Fragment key={protocol.id}>
-                    <RecentRunCard
+                    <RecentProtocolRunCard
                       lastRun={lastRun}
                       protocolId={protocolId}
                       protocolName={
@@ -84,112 +80,120 @@ export function RobotDashboard(): JSX.Element {
   )
 }
 
-interface RecentRunCardProps {
-  protocolName: string
-  protocolId: string
-  lastRun?: string
-}
+// interface RecentRunProtocolCardProps {
+//   protocolName: string
+//   protocolId: string
+//   lastRun?: string
+// }
 
-function RecentRunCard(props: RecentRunCardProps): JSX.Element {
-  const { t, i18n } = useTranslation('device_details')
-  const { protocolName, protocolId, lastRun } = props
-  const missingProtocolHardware = useMissingProtocolHardware(protocolId)
-  const history = useHistory()
-  const isSuccess = missingProtocolHardware.length === 0
+// function RecentRunProtocolCard(props: RecentRunProtocolCardProps): JSX.Element {
+//   const { t, i18n } = useTranslation('device_details')
+//   const { protocolName, protocolId, lastRun } = props
+//   const missingProtocolHardware = useMissingProtocolHardware(protocolId)
+//   const history = useHistory()
+//   const isSuccess = missingProtocolHardware.length === 0
 
-  const CARD_STYLE = css`
-    &:active {
-      background-color: ${isSuccess
-        ? COLORS.green_three_pressed
-        : COLORS.yellow_three_pressed};
-    }
-    &:focus-visible {
-      box-shadow: 0 0 0 ${SPACING.spacing1} ${COLORS.fundamentalsFocus};
-    }
-  `
+//   const CARD_STYLE = css`
+//     &:active {
+//       background-color: ${isSuccess
+//         ? COLORS.green_three_pressed
+//         : COLORS.yellow_three_pressed};
+//     }
+//     &:focus-visible {
+//       box-shadow: 0 0 0 ${SPACING.spacing1} ${COLORS.fundamentalsFocus};
+//     }
+//   `
 
-  const missingProtocolHardwareType = missingProtocolHardware.map(
-    hardware => hardware.hardwareType
-  )
-  const missingProtocolPipetteType = missingProtocolHardwareType.filter(
-    type => type === 'pipette'
-  )
-  const missingProtocolModuleType = missingProtocolHardwareType.filter(
-    type => type === 'module'
-  )
+//   const PROTOCOL_TEXT_STYLE = css`
+//     display: -webkit-box;
+//     -webkit-box-orient: vertical;
+//     -webkit-line-clamp: 5;
+//     overflow: hidden;
+//   `
 
-  let chipText: string = t('ready_to_run')
-  if (
-    missingProtocolPipetteType.length === 0 &&
-    missingProtocolModuleType.length > 0
-  ) {
-    chipText = t('missing_module', {
-      num: missingProtocolModuleType.length,
-      count: missingProtocolModuleType.length,
-    })
-  } else if (
-    missingProtocolPipetteType.length > 0 &&
-    missingProtocolModuleType.length === 0
-  ) {
-    chipText = t('missing_pipette', {
-      num: missingProtocolPipetteType.length,
-      count: missingProtocolPipetteType.length,
-    })
-  } else if (
-    missingProtocolPipetteType.length > 0 &&
-    missingProtocolModuleType.length > 0
-  ) {
-    chipText = t('missing_both', {
-      numMod: missingProtocolModuleType.length,
-      numPip: missingProtocolPipetteType.length,
-    })
-  }
-  const handleCardClick = (): void => {
-    history.push(`protocols/${protocolId}`)
-  }
+//   const missingProtocolHardwareType = missingProtocolHardware.map(
+//     hardware => hardware.hardwareType
+//   )
+//   const missingProtocolPipetteType = missingProtocolHardwareType.filter(
+//     type => type === 'pipette'
+//   )
+//   const missingProtocolModuleType = missingProtocolHardwareType.filter(
+//     type => type === 'module'
+//   )
 
-  return (
-    <Flex
-      aria-label="RecentRunCard"
-      css={CARD_STYLE}
-      flexDirection={DIRECTION_COLUMN}
-      padding={SPACING.spacing5}
-      gridGap={SPACING.spacing5}
-      backgroundColor={isSuccess ? COLORS.green_three : COLORS.yellow_three}
-      width="25.8125rem"
-      borderRadius={BORDERS.size_four}
-      onClick={handleCardClick}
-    >
-      {/* marginLeft is needed to cancel chip's padding */}
-      <Flex marginLeft={`-${SPACING.spacing4}`}>
-        <Chip
-          type={isSuccess ? 'success' : 'warning'}
-          background={false}
-          text={i18n.format(chipText, 'capitalize')}
-        />
-      </Flex>
-      <Flex width="100%" height="14rem">
-        <StyledText
-          fontSize={TYPOGRAPHY.fontSize32}
-          fontWeight={TYPOGRAPHY.fontWeightLevel2_bold}
-          lineHeight={TYPOGRAPHY.lineHeight42}
-        >
-          {protocolName}
-        </StyledText>
-      </Flex>
-      <StyledText
-        fontSize={TYPOGRAPHY.fontSize22}
-        fontWeight={TYPOGRAPHY.fontWeightRegular}
-        lineHeight={TYPOGRAPHY.lineHeight28}
-        color={COLORS.darkBlack_seventy}
-      >
-        {i18n.format(t('last_run_time'), 'capitalize')}{' '}
-        {lastRun != null
-          ? formatDistance(new Date(lastRun), new Date(), {
-              addSuffix: true,
-            }).replace('about ', '')
-          : ''}
-      </StyledText>
-    </Flex>
-  )
-}
+//   let chipText: string = t('ready_to_run')
+//   if (
+//     missingProtocolPipetteType.length === 0 &&
+//     missingProtocolModuleType.length > 0
+//   ) {
+//     chipText = t('missing_module', {
+//       num: missingProtocolModuleType.length,
+//       count: missingProtocolModuleType.length,
+//     })
+//   } else if (
+//     missingProtocolPipetteType.length > 0 &&
+//     missingProtocolModuleType.length === 0
+//   ) {
+//     chipText = t('missing_pipette', {
+//       num: missingProtocolPipetteType.length,
+//       count: missingProtocolPipetteType.length,
+//     })
+//   } else if (
+//     missingProtocolPipetteType.length > 0 &&
+//     missingProtocolModuleType.length > 0
+//   ) {
+//     chipText = t('missing_both', {
+//       numMod: missingProtocolModuleType.length,
+//       numPip: missingProtocolPipetteType.length,
+//     })
+//   }
+//   const handleCardClick = (): void => {
+//     history.push(`protocols/${protocolId}`)
+//   }
+
+//   return (
+//     <Flex
+//       aria-label="RecentRunCard"
+//       css={CARD_STYLE}
+//       flexDirection={DIRECTION_COLUMN}
+//       padding={SPACING.spacing5}
+//       gridGap={SPACING.spacing5}
+//       backgroundColor={isSuccess ? COLORS.green_three : COLORS.yellow_three}
+//       width="25.8125rem"
+//       borderRadius={BORDERS.size_four}
+//       onClick={handleCardClick}
+//     >
+//       {/* marginLeft is needed to cancel chip's padding */}
+//       <Flex marginLeft={`-${SPACING.spacing4}`}>
+//         <Chip
+//           type={isSuccess ? 'success' : 'warning'}
+//           background={false}
+//           text={i18n.format(chipText, 'capitalize')}
+//         />
+//       </Flex>
+//       <Flex width="100%" height="14rem">
+//         <StyledText
+//           fontSize={TYPOGRAPHY.fontSize32}
+//           fontWeight={TYPOGRAPHY.fontWeightLevel2_bold}
+//           lineHeight={TYPOGRAPHY.lineHeight42}
+//           css={PROTOCOL_TEXT_STYLE}
+//         >
+//           {protocolName}
+//         </StyledText>
+//       </Flex>
+//       <StyledText
+//         fontSize={TYPOGRAPHY.fontSize22}
+//         fontWeight={TYPOGRAPHY.fontWeightRegular}
+//         lineHeight={TYPOGRAPHY.lineHeight28}
+//         color={COLORS.darkBlack_seventy}
+//       >
+//         {i18n.format(t('last_run_time'), 'capitalize')}{' '}
+//         {lastRun != null
+//           ? formatDistance(new Date(lastRun), new Date(), {
+//               addSuffix: true,
+//             }).replace('about ', '')
+//           : ''}
+//       </StyledText>
+//     </Flex>
+//   )
+// }
