+++ app/src/pages/Protocols/hooks/index.ts
@@ -90,7 +90,5 @@ export const useMissingProtocolHardware = (
   protocolId: string
 ): ProtocolHardware[] => {
   const requiredProtocolHardware = useRequiredProtocolHardware(protocolId)
-  return requiredProtocolHardware.filter(
-    hardware => hardware.connected === false
-  )
+  return requiredProtocolHardware.filter(hardware => !hardware.connected)
 }
